Relation Deletion Conflict Resolution
-Commit #39 had deletion of relations pointing to deleted boxes done in
the boxes themselves. At the same time I had implemented a way where the
relations were instead deleted in the controller, which keeps the boxes
from having to keep track of relations too. Both worked but for the sake
of coupling and following our MVC structure I thought it be best we go
with the relations being deleted in the controller.